requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.2
werkzeug==2.3.7
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
librosa==0.10.1
soundfile==0.12.1
gunicorn==21.2.0
google-auth==2.23.0
google-auth-oauthlib>=0.4.1,<0.5
requests==2.31.0 